requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
pdfplumber==0.9.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.1
